(T): make sure that parse-fail test actually fail
